Move context processor into filters blueprint
`create_magnet_from_es_info` from nyaa.torrents
