+++ nyaa/template_utils.py
@@ -1,15 +1,38 @@
 import os.path
+from base64 import b32encode
 from datetime import datetime
 from email.utils import formatdate
+from urllib.parse import urlencode
 
 import flask
 from werkzeug.urls import url_encode
 
 from nyaa import app
 from nyaa.backend import get_category_id_map
+from nyaa.torrents import get_default_trackers
 
 bp = flask.Blueprint('template-utils', __name__)
 _static_cache = {}  # For static_cachebuster
+
+
+# ######################## CONTEXT PROCESSORS ########################
+
+# For processing ES links
+@bp.app_context_processor
+def create_magnet_from_es_info():
+    def _create_magnet_from_es_info(display_name, info_hash, max_trackers=5, trackers=None):
+        if trackers is None:
+            trackers = get_default_trackers()
+
+        magnet_parts = [
+            ('dn', display_name)
+        ]
+        for tracker in trackers[:max_trackers]:
+            magnet_parts.append(('tr', tracker))
+
+        b32_info_hash = b32encode(bytes.fromhex(info_hash)).decode('utf-8')
+        return 'magnet:?xt=urn:btih:' + b32_info_hash + '&' + urlencode(magnet_parts)
+    return dict(create_magnet_from_es_info=_create_magnet_from_es_info)
 
 
 # ######################### TEMPLATE GLOBALS #########################
