+++ nyaa/torrents.py
@@ -85,24 +85,6 @@ def create_magnet(torrent, max_trackers=5, trackers=None):
     return 'magnet:?xt=urn:btih:' + b32_info_hash + '&' + urlencode(magnet_parts)
 
 
-# For processing ES links
-@app.context_processor
-def create_magnet_from_es_info():
-    def _create_magnet_from_es_info(display_name, info_hash, max_trackers=5, trackers=None):
-        if trackers is None:
-            trackers = get_default_trackers()
-
-        magnet_parts = [
-            ('dn', display_name)
-        ]
-        for tracker in trackers[:max_trackers]:
-            magnet_parts.append(('tr', tracker))
-
-        b32_info_hash = base64.b32encode(bytes.fromhex(info_hash)).decode('utf-8')
-        return 'magnet:?xt=urn:btih:' + b32_info_hash + '&' + urlencode(magnet_parts)
-    return dict(create_magnet_from_es_info=_create_magnet_from_es_info)
-
-
 def create_default_metadata_base(torrent, trackers=None):
     if trackers is None:
         trackers = get_trackers(torrent)
